+++ multimodal/src/autogluon/multimodal/utils/__init__.py
@@ -8,6 +8,7 @@ from .config import (
     get_local_pretrained_config_paths,
     get_pretrain_configs_dir,
     parse_dotlist_conf,
+    save_timm_config,
     update_config_by_rules,
     update_tabular_config_by_resources,
 )
@@ -43,14 +44,7 @@ from .matcher import compute_semantic_similarity, convert_data_for_ranking, crea
 from .metric import compute_ranking_score, compute_score, get_minmax_mode, get_stopping_threshold, infer_metrics
 from .misc import logits_to_prob, merge_bio_format, shopee_dataset, tensor_to_ndarray, visualize_ner
 from .mmcv import CollateMMDet, CollateMMOcr, send_datacontainers_to_device, unpack_datacontainers
-from .model import (
-    create_fusion_model,
-    create_model,
-    list_timm_models,
-    modify_duplicate_model_names,
-    save_timm_config,
-    select_model,
-)
+from .model import create_fusion_model, create_model, list_timm_models, modify_duplicate_model_names, select_model
 from .object_detection import (
     COCODataset,
     bbox_xyxy_to_xywh,
