+++ multimodal/src/autogluon/multimodal/utils/config.py
@@ -1,4 +1,5 @@
 import copy
+import json
 import logging
 import os
 import warnings
@@ -20,6 +21,7 @@ from ..constants import (
     REGRESSION,
     VALID_CONFIG_KEYS,
 )
+from ..models import TimmAutoModelForImagePrediction
 from ..presets import get_automm_presets, get_basic_automm_config, get_preset_str
 
 logger = logging.getLogger(AUTOMM)
@@ -560,3 +562,47 @@ def get_pretrain_configs_dir(subfolder: Optional[str] = None):
     if subfolder:
         pretrain_config_dir = os.path.join(pretrain_config_dir, subfolder)
     return pretrain_config_dir
+
+
+def _filter_timm_pretrained_cfg(cfg, remove_source=False, remove_null=True):
+    filtered_cfg = {}
+    keep_null = {"pool_size", "first_conv", "classifier"}  # always keep these keys, even if none
+    for k, v in cfg.items():
+        if remove_source and k in {"url", "file", "hf_hub_id", "hf_hub_id", "hf_hub_filename", "source"}:
+            continue
+        if remove_null and v is None and k not in keep_null:
+            continue
+        filtered_cfg[k] = v
+    return filtered_cfg
+
+
+def save_timm_config(
+    model: TimmAutoModelForImagePrediction,
+    config_path: str,
+):
+    """
+    Save TIMM image model configs to a local file.
+
+    Parameters
+    ----------
+    model
+        A TimmAutoModelForImagePrediction model object.
+    config_path:
+        A file to where the config is written to.
+    """
+    config = {}
+    pretrained_cfg = _filter_timm_pretrained_cfg(model.config, remove_source=True, remove_null=True)
+    # set some values at root config level
+    config["architecture"] = pretrained_cfg.pop("architecture")
+    config["num_classes"] = model.num_classes
+    config["num_features"] = model.out_features
+
+    global_pool_type = getattr(model, "global_pool", None)
+    if isinstance(global_pool_type, str) and global_pool_type:
+        config["global_pool"] = global_pool_type
+
+    config["pretrained_cfg"] = pretrained_cfg
+
+    with open(config_path, "w") as f:
+        json.dump(config, f, indent=2)
+        logger.info(f"Timm config saved to {config_path}.")
